requirements.txt
discord.py~=1.6.0
setuptools~=54.2.0
build==0.3.1.post1